requirements.txt
fastapi[standard]
uvicorn
google-generativeai
python-dotenv
aiomysql
bcrypt # for password hashing
pydantic[email]
PyJWT # for JWT authentication
google-genai # Updated version of google-generativeai